4.21*: blogilistan laajennus, step9 DONE, (tried userExtractor and failed miserably)

+++ controllers/blogs.js
@@ -3,24 +3,22 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
+/*
 const getTokenFrom = request => {
   const authorization = request.get('authorization')
   if (authorization && authorization.startsWith('Bearer ')) {
     return authorization.replace('Bearer ', '')
   }
   return null
-}
+}*/
 
 blogsRouter.post('/', async (request, response) => {
   console.log('BLOGSROUTER')
   const body = request.body
-
-  //console.log('BODY', body)
-  const token = getTokenFrom(request)
-  //console.log('TOKEN', token)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  //console.log('decodedToken', decodedToken)
-  if (!decodedToken.id) {
+  //const user = request.user
+  //console.log('USERTOKEN', user)
+  const userOfToken = jwt.verify(request.token, process.env.SECRET)
+  if (!userOfToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
   if (!body.title || !body.url) {
@@ -28,7 +26,9 @@ blogsRouter.post('/', async (request, response) => {
   }
   //console.log('BODY', body)
   //const user = await User.findById(body.userId)
-  const user = await User.findById(decodedToken.id)
+  
+  const user = await User.findById(userOfToken.id)
+  
   //console.log('USER', user)
   //const users = await User.findById({})
   const blog = new Blog({
@@ -45,6 +45,7 @@ blogsRouter.post('/', async (request, response) => {
     blog.likes = body.likes
   }
 
+  console.log('BLOG', blog)
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
@@ -73,6 +74,22 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const userOfToken = jwt.verify(request.token, process.env.SECRET)
+  console.log('DECODED', userOfToken)
+  if (!userOfToken.id || !request.token) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'Blog not found' })
+  }
+
+  if (blog.user.toString() !== userOfToken.id.toString()) {
+    return response.status(403).json({ error: 'You do not have rights to delete' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
